Only send destroy if the context understands it.
SVN 1.4.6, among others, doesn't.

+++ right_scraper_svn/lib/right_scraper_svn/svn_client.rb
@@ -67,7 +67,7 @@ module RightScale
       begin
         yield context
       ensure
-        context.destroy
+        context.destroy if context.respond_to?(:destroy)
       end
     end
   end
